requirements.txt
python-dotenv
langchain
langgraph
langchain-groq
langchain-core
langchain-community
langchain-huggingface
sentence-transformers
faiss-cpu
typing-extensions
pypdf
pymupdf
langchain-text-splitters
requests

streamlit
langchain
langchain-core
langchain-community
langchain-huggingface
langchain-groq
langgraph
faiss-cpu
python-dotenv